[sysmem] Add per-plane info to format
For multi-plane formats the byte offset of some planes is flexible and
must be specified somehow. For example, the amlogic MPEG2 decoder always
has an offset for the UV plane of 1920*1152, while the h.264 decoder
puts the data right after the Y plane.

The stride could also vary per plane, so put that in the ImagePlane
struct.

Test: build

Change-Id: I8c621682272caa26bc5633f42d3e941586dab4b8