[escher] Separate mesh vertex positions from other attributes.
This follows the recommendation from the ARM Mali best practices guide,
as an Early-Z optimization (the vertex shader is run in two stages,
where the instructions required to compute the screen-space position
are run first, and the rest is run only if Early-Z is passed).

However, the main motivation is that this separation makes geometric
mesh algorithms easier to write.  For example, when splitting an edge
the positions can be used separately to determine an interpolation
parameter, when can then be used to generically lerp the rest of the
attributes.

To verify that this works, we modify the rounded rectangle tessellator
so that it can generate both interleaved/separated uv/pos attributes,
and also modify PaperRenderFuncs to handle both.

ES-131 #done

TEST: escher_unittests, waterfall/waterfall2

Change-Id: Ied71c769aa277aeba085ac920e6e2cc80b3b0f05