[cobalt] Switch back to using integer metric IDs in the FIDL API.
The user will still register Metrics in the config file using only
a Metric name. He will not have to write a metric ID. The Config
Parser will generate a metric ID by hashing.

We have several choices for how the user will get the integer ID in
order to pass it in the API including a web page that shows them
the ID for a name and having the Config Parser generate a header
file with a generated const.

TEST=compiles

Change-Id: Iaa2233dd1e0dc03f7a9cd24994d864538de2ded0