[camera] Adding eventpair to allow stream shutdown
The camera manager will operate by passing a camera::Stream
interface directly to applications.  This addition to the camera
interface (which is hidden from applications using the camera
manager) allows the camera manager to retain a handle to the
stream that was created, so it can signal the camera driver
to kill a stream if needed.

Currently, the implementation only considers signalling from the
camera manager to the driver.  Future CLs will allow the camera to
close the eventpair to signal to the camera manager that it has freed
a resource.

TEST = CQ

Change-Id: I71411b80ac0e97e3854da3fec5ced2dabda665c3