[wlanstack2] Report RSSI stats from wlanstack2.
This CL completes the migration of telemetry from MLME to wlanstack2.

The reporting is done is such a way that the stats are not reset
after each report. A cache with last reported stats is used to only send
to Cobalt the stats from the last period of time.

NET-1343 #done
NET-1239

Test: Build, boot and check Cobalt acceptance.

Change-Id: Ie508be82ee6b36b68b1e1ff815067c2fa033e869