[magma] Add functions to decode fuchsia.sysmem.SingleBufferSettings
We'll soon want to be able to pass fuchsia.sysmem.SingleBufferSettings
into vulkan drivers, and they need to be able to read information about
the format. By adding magma functions we can allow that without causing
them to depend on the precise layout of the structure.

This takes in an encoded fidl message instead of a
fuchsia_sysmem_SingleBufferSettings struct because
fuchsia.sysmem.SingleBufferSettings will eventually change to being a
fidl table, and we want to use the wire format so it's as compatible as
possible.

Test: vim2:go/magma-tps#L0

Change-Id: Ie5b417761f61d22d49cd65a5e04c7bfe83b67053