[intents] Transform intents prior to handing to IntentHandler
The framework will now transform the intent which is passed to the
IntentHandler by converting Intent.parameters which are either
link_name or link_path to a link_name parameter in the handler's
namespace.

This makes all the different intent parameter types sensible to the
handler.

Ideally this would be more sophisticated to avoid overwriting link data
when two or more consecutive intents with overlapping parameter names
are handled. However the need for such functionality will be removed in
a future rework of links, and is therefore left unimplemented.

MI4-1248 #done

TESTED=tests/intents

Change-Id: I2144f8b4766df56ba742aefc6742ed01f4f7a836