[kernel][debug] Variant of _panic() for pre-formatted panic strings.
ZX-2527 describes a call graph cycle in the panic handler caused
by the call to printf in _panic() calling _panic() recursively.
Adding a _panic() variant that avoids calling printf() that can
be used from __stack_chk_fail() and elsewhere. This commit adds
a variant of panic() where the panic message is pre-formatted.
The new call also omits the call location.

Bug: ZX-2527

Test: Verified by Roland McGrath using instrumented builds.

Change-Id: Iabd5769d40f633db98269859f9000153db14b180