[platform-bus] Remove proxying of USB mode switch protocol
This protocol is only being used for Hikey960, which needs to change some GPIOs
when switching USB mode. But now that we have support for child platform devices,
this logic can be pushed out of the board driver and into a platform device driver
which has the dwc3 driver as a child.

TEST: manual testing on hikey960 and imx8m-evk
Change-Id: Ia3c1b9540d0813d3d873f22007e3e22efb219975