[media][tools] Update signal_generator
The signal_generator tool has been updated to generate
its audio signal on-the-fly for each packet, instead of
(previously) generating a second of audio ahead-of-time
that is then looped.

The signal_generator tool accepts a number of cmd-line
parameters to set the signal type, frequency, duration,
buffer size, and more.  This CL converts the frequency and
duration parameters from integers to doubles, to handle
arbitrary values. Duration can be as small as 0, and
frequencies as low as 0.25 have been validated.

Also, the packet-size parameter has been changed from
milliseconds-per-payload value to frames-per-payload,
making this control much more precise.

Finally, this CL enables signal_generator to change the
System Mute setting, in addition to its existing
stream-specific and System Gain settings.

Test: manual and CQ

Change-Id: Iebf0fe8fabef7a04bafb78ec28791042f1650b31