[wlanif] Introduce wlanif (fullmac driver)
The wlanif driver provides a single ddk driver for fullmac devices
(which publish as wlanif-impl devices), with interfaces for both
MLME messages and data (ethernet) messages.
As a first step, the MLME interface is presented as a direct
translation to and from FIDL MLME messages, but this is expected
to change as functionality moves into and out of drivers, and as
the wlanmac driver is migrated to being a wlanif-impl device.

NET-1189

Test: Implemented a dummy wlanif-impl device driver which simulated
ethernet and wlan queries and MLME scanning.

Change-Id: I30f4f46f32020b4c3fe0784e85f7077769e52817