[services] StartupContext lazily initializes services.
Presently, StartupContext initializes both the launcher and the
environment handles when it is constructed. This means that if the
component did not include Environment or Launcher in its sandbox,
appmgr will print a WARNING message even if the component never actually
attempts to use them. We change this to initialize these handles lazily,
when they are accessed.

In a future change we should probably initialize these channels with
Lookup instead of ConnectToService, so a helpful error will be returned
if the component does not have access to the service (CP-124).

TESTED=no externally visible change in behavior, so no additional tests

Change-Id: I1c044b39febe1564312399a0ff60b441f3e23884