[audio] Consolidate mixer bookkeeping
Both the StandardOutputBase and the AudioInImpl classes have the
concept of a mixer bookkeeping object that contains a pointer to
the actual mixer as well as the parameters that control mixer rate
calculations, such as the local-to-source clock, the dest-to-source
clock, and the step_size (and modulo/denominator).

This CL takes the first step toward eventually relocating these
bookkeeping objects, from StandardOutputBase and AudioInImpl into
the Mixer class itself. In this first step, we unify the two
bookkeeping implementations. Specifically, this CL eliminates
the StandardOutputBase::AudioOutBookkeeping and
AudioInImpl::CaptureLinkBookkeeping subclasses, in favor of a single
unified AudioLink::Bookkeeping class that is used by both
StandardOutputBase and AudioInImpl.

Test: manual and CQ

MTWN-129 #comment

Change-Id: I08c17af7798e4b913f4ecadc240eddc8c1e3b3bc