[bt] Implement Host.Connect, use for connections in BT-GAP
Add the Host.Connect fidl method and implement in the host server.
Only connect LE if device in cache and supports LE.
Update bt-gap to connect via the Host's method.

NET-1092 #done

Test: manual - Connect to a remote device by using bt-cli's connect command,
and ensure that the remote device state changes to connected.

(NET-2024 has been filed to add an integration test that exercises this
method)

Change-Id: I84554350b14fa665334420b8319ae6ebc4d209e5