[zircon][sdio] Use single byte command for shorter transfers
Use CMD52 for transfers that are less than 4 bytes. The
CMD53 uses ping pong buffers in aml-sd-emmc controller in the case of PIO mode.
From amlogic documentation it feels to me that the register space should be accessed
4 bytes at a time. So lets use multiple CMD52 instead of CMD53 for transfers less than
size 4 to avoid complications and controller errors. Otherwise, the write silently succeeds
but it does not write the actual data.

Test: Tested that the block size is written correctly on vim2 after the
change. Prior to this, the block sizes are wrong.

Change-Id: Ie8fa0bd38b2f2a3354e197eceafafcd13cd00a1f