[debugger] Add thread controller system.
Replaces the "RunUntil" functions with a ThreadController. Currently
only the "until" command is hooke up with this system, and a per-thread
"finish" and a process-wide "until" are broken. These will be fixed in a
follow-up.

BreakpointControllers have been removed, along with all of the
infrastructure associated with notifying a controller about a breakpoint
hit and acting on that. Moving forward, this will all be done with
ThreadControllers so there is one place to check for what should happen
on a stop (previously there was a bug if the user did a "until" to the
same address of a user breakpoint).

TEST=will do in followup when "finish" works.

Change-Id: I871db7ae587dabac0549b26c2d1524cfc7cc7c81