[zircon][sdio] Restrict the request sizes in pio mode to be 4-byte aligned
amlogic controller uses 2 512-byte buffers in mmio region for
transfering data in PIO mode. From amlogic documentation, these buffers
are 32 bit accessible. Copying byte by byte, looks like it succeeds but
the actual data does not get copied. For this reason, the request sizes
should be a multiple of 4 in PIO mode. For now, the restriction is also
added to dma transfers in sdio_rw_data. The reason for this is that, although
the upper layer wants to use dma, the astro board does not support dma yet,
and request is converted to PIO mode. Along with that, this change also
removes the unnecessary cache ops as the mmio region is mapped with
uncached policy.

Test: Tested that sdio probe was successful on vim2. Tested 16 byte read
with and without dma locally in my sdio code.

ZX-2548 #comment:Fix the write failures in PIO mode.

Change-Id: I7f4b0e93dc6ebb446692d23a9add6b7c1b27abf0