[virtcon] Avoid reading from WC memory
Reading from memory with a cache policy other than CACHED is
potentially very slow. Since gfx_surface needs to read from its buffer
for scrolling, if the framebuffer vmo can't be mapped CACHED then use a
second gfx_surface for rendering and blit the result to the framebuffer
vmo.

This change relies on the display driver setting an appropriate cache
policy with zx_set_cache_policy and keeping its own mapping of the VMO
to prevent the virtcon from setting the cache policy.

Bug: ZX-2198
Test: Check virtcon scrolling speed on NUC with intel_i915 disabled
Change-Id: I8af71b72ed9fff720947e685dee286d9c0edaf26