[cobalt-client]: Update for cobalt2.fidl.
Update internal data to use a representation compatible with
cobalt2.fidl. This is, no ObservationPart or Value, and dimensions
are represented as a pair of ints, depicting an event type and
an index within the given event type.

Until multiple dimensions are supported, we can assume one index being
passed.

TESTS=cobalt-client-tests

Change-Id: Ia4afb4015c38aeeda4152a043c8997e89f4d6efe