[debugger] Add a "finish" thread controller.
Hooks it up to the "finish" command.

Improves ThreadController infrastructure to allow a ThreadController to
fail in setup. This fixes the problem of doing
  until NonexistantSymbol
and having the thread transparently continue.

The thread registration is also cleaned up so that that the thread
doesn't need to be passed to the controller constructor.

Moves the test infrastructure from thread_impl_unittest.cc to a separate
file (thread_impl_test_support.*) so it can be shared by thread stepping
tests. This code is unchanged.

TEST=unit

Change-Id: Ib1af5b93aa328f5b8a83d5b2539d6f4985a33b71