[dev][platform-bus] separate regular metadata from bootloader metadata
In the platform bus protocol, we use the pbus_metadata_t struct
for board drivers to pass metadata to the platform devices.
Instead of overloading this struct to work both for immediately
provided metadata and metadata passed from the bootloader, we now have
a separate struct pbus_boot_metadata_t used for specifying bootloader
metadata to be assigned to a platform device.

ZX-2513 #in progress

TEST: booted VIM2, astro and gauss and verified that drivers that
      use metadata started up successfully.

Change-Id: I7065fd497646f386987cb5dc317d1c66daa8df0a