[scenic][input] Scenic input subsystem: Init fix.
The initialization needs to wait for GFX, but also register early enough
to be available for CommandDispatcher creation; otherwise, the input
subsystem's CommandDispatcher never gets created.

The fix is to register the input subsystem promptly in
bin/ui/scenic/app.cc, but have its constructor add a SetToInitialized()
callback in the GFX post-init subsystem.

TEST=run hello_input

Change-Id: I35fe673a1f96c0bdd300e5ba2163a67ada996fd6