[region-alloc] Remove custom new/delete
Doing some cleanup of my more ancient bug backlog and I ran across
this one.  Follow the bug link for all of the technical details; but
the TL;DR is to remove any need for custom new/delete behavior and to
move to the standard zircon AllocChecker approach.

ZX-334 #done

Test: build and region-alloc unit tests
Change-Id: I663232d0471d2adeb2f0d6cf33bd2211a1693e96