[nand-util] Adds a debugging tool for nand drivers.
This is a user mode tool that uses the nand-broker driver to access a
nand device and performs basic operations.

The operations supported at this point are:
Display basic nand info.
Display data from the chip "page0".
Decodes the bad block table(s), to show the list of bad blocks.
Show the contents of any page.
Erase a block.

The app can work on top of any nand protocol driver (with the broker),
but the binding rules have to be tweaked as needed to cause the broker
to load where needed.

Test: Functionality verified on top of the standard nand device.
Change-Id: I52f135b23484081bf14e56318f21f431926277b6