[inspect] Fix flake in inspect test.
We do not currently have a way to determine if components have started
and are ready, so we need to depend on side effects becoming visible.
The underlying issue will be addressed in the future.

For now, check that the exported directory exists to fix flakes in this
test.

CP-130: #comment
TEST=run_test_component inspect_integration_tests
Change-Id: Iee63dac51e20a46a8c7dff4b30e7269de162537a

Change-Id: I1e1b2e0ec587ae9767215b55449bc04f03de0311