[component_context] Expose ComponentContext directly
This allows client wrappers which depend on functionality in
ComponentContext to connect to it directly. Previously such a client
library would need to determine whether it was an agent or a module, get
the appropriate context, and then get the ComponentContext from there.

ModuleContext and AgentContext still expose the ComponentContext.

TESTED=No change in functionality.

Change-Id: I56fed8b842868c0edeba9c8e57d69df7574bb7ba