[kernel][object] SafeDeleter: Fix potential stack overflow on OOM
Allocate SafeDeleter on the stack instead of on the heap.

Before, the SafeDeleter is allocated on the heap and kept around until
the thread exits.

That creates the possibility that the allocation fails, in which case
the code would fall back to doing an unsafe deletion of complex
objects which could overflow the kernel stack.  (To test this, I
changed SafeDeleter::Get() to return null.  With that change,
channel-test causes a kernel panic.)

Instead, we can allocate the SafeDeleter on the stack and register it
in TLS only for the duration of the compound deletion.

Bug: none
Test: channel-test (see above)
Change-Id: If69881c77b3726f8fc2361210e35be88425b09f4