[debugger] Move "step" to ThreaControllers.
Removes the Step() function from Thread and implements this
functionality in terms of ThreadController.

There are two controllers added, one which steps given a range, and one
which does exactly "StepInto". The latter converts the current line into
a range supplied to the former. I did this split because I expect the
range stepper will be used as a component of other steppers in the
future.

Add a capability for a ThreadController to specify how the thread should
be run while that controller is active. Don't try to continue threads
that had non-debug exceptions (test added for this case).

TEST=manual

Change-Id: Ie7d084037ac3d9f5f2f81514d29c701389894930