[catapult] Reject input not of type array
catapult_converter expects input of type array, and thus should reject
input where that is not the case.

Previously, it was possible for non-array input to get silently
converted to an empty array.  Note that these would have been caught by
debug assertions in rapidjson, however we run with those disabled.

Test:
$ pwd
/tmp
$ echo '[]' > array.json
$ echo 'null' > null.json
$ /pkgfs/packages/catapult_converter/0/bin/app --input=array.json --log-url=https://fakeurl.com --bots=none --masters=none --execution-timestamp-ms=1
$ /pkgfs/packages/catapult_converter/0/bin/app --input=null.json --log-url=https://fakeurl.com --bots=none --masters=none --execution-timestamp-ms=1
Expected input document to be of type array, and got null instead

Change-Id: Ie9477e4d3ba879ff3155d46427f31120c4359d50