[fs][minfs][blobfs] Simplify BlockTxn.
This patch acts as a pre-requisite for latter patches
which require transactions that combine operations, such
as "BLOCKIO_WRITE" and "BLOCKIO_FLUSH".

- Remove templatization of BlockTxn. This will allow
the implementation to be moved into a ".cpp" file
in a future patch. ReadTxn and WriteTxn still exist
as simple type-safe wrappers around an untemplated
BlockTxn.
- Rename "Flush" to "Transact"; "Flush" is an overloaded
term in the prescence of "BLOCKIO_FLUSH".
- Define an explicit TransactionHandler pure virtual
interface, to explicitly define the usage of a BlockTxn,
rather than relying on implicit templatization rules.

Test: runtests, fs-tests

Change-Id: Ic18687d31b44876ea9c06b49256ab25bd56f75db