[devmgr] Stop printing error message that is now triggered by platform bus
A recent change in the platform bus driver is now binding the platform-bus.proxy.so
driver with ZX_PROTOCOL_PLATFORM_PROXY for protocl proxy client drivers.
This is working fine, but the devmgr is complaining that the driver does not have
a "bind" callback:

[03060.767] 03030.03096> devhost[PlatformProxy] bind driver '/boot/driver/platform-bus.proxy.so' failed: -2
[03060.767] 01044.01047> devcoord: rpc: bind-driver '04:02:1' status -2

This is normal operation though, since the driver is using the create callback instead.
This CL adjusts the error reporting in the devhost DC_OP_BIND_DRIVER callback to handle this case better.

TEST: boot on VIM2, error message went away.
Change-Id: Ie14a0986dd2737bab76567f477e1a36a8de7fd40