[debugger] Add step over ("next") command.
This adds a StepOverThreadController object which implements "next" and
"nexti" in the frontend. This currently doesn't quite work right due to
bugs in the symbol line handling. Those will be addressed separately in
a followup.

The algorithm is described in step_over_thread_controller.h

Includes some new testing infrastructure for testing thread controllers.

TEST=unit

Change-Id: Id61ee173e5c46fceb68246e5b8b7362e5605d0d8