[network_time] Add integration tests for network_time service.
- Allow the network_time service take an optional config_path from which
to load a Roughtime server list
- Introduce LocalRoughTimeServer, wrapping roughtime::SimpleServer, for
serving constant time values in tests.
- Add an integration test that starts a local Roughtime server and the
network_time service and confirms that the system clock is updated to
the values provided by the server.
- Add an integration test that smoke-tests the network_time service
against the production list of Roughtime servers (non-hermetic). [DISABLED
until it can be set up to run in CI.]
- Some cleanup in TimeServerConfigTest.

TEST=run_test_component system_time_updater_test
CP-131: #progress

Change-Id: I21b3e73c49d37beee4cb57b4d6b909f3e99f5d05