[garnet][sdio] Move the wifi gpio handling to where it belongs.
out-of-bandwidth interrupt is gpio driver interrupt which is specific to
sdio client drivers(wifi). Prior to Mike's recent platform bus changes, it
was not possible for child drivers to configure gpios that are passed
via board driver. Now that we have that capability, lets move the gpio
configure to wifi driver where it belongs. This will avoid all the
unnecessary protocol ops to get the oob interrupt configured. This will
also help check for spurious interrupts. This changeset also makes use
of the board specific wifi information by retreiving the metadata from
the board driver.

Test: Called sdio_intr_register and verified that the interrupt was
configured and interrupts are received.

Change-Id: I926a26b4f18cde7042137c51a8e2ec0db81fe464