[benchmarks] Support the new schema in trace-based benchmarks.
Adopt the use of "split_first" parameter in spec files for the
trace-based benchmarks, as well as exporting the results using
the new (flattened) schema.

As a first step, the results will only be exported using the new schema
(defined at docs/development/benchmarking/results_schema.md) when the
split_first parameter is explicitly set to true. Once we make sure the
results are rendered correctly, we can switch all the trace-based
benchmark to it and remove the old schema.

TEST=pkgfs/packages/trace_tests/0/test/trace_tests

LE-590, IN-568

Change-Id: Iae6f70ab1e2343fb018a19ed78c21dc0ba71dd9c