[fidl] Adding missing dependency to fuchsia.ui.views1 fidl library
A recent fidlc change (https://fuchsia-review.googlesource.com/c/zircon/+/185902) led to breaking the garnet roller (https://fuchsia-review.googlesource.com/c/peridot/+/190496). This adds the missing dependency, and was compiled with the new fidlc changes to confirm. Without this import, the error seen in the build is exactly reproduced with the Test commands below.

Test: fx set arm64 && fx build peridot/bin/user_runner/story_runner:link_impl_unittest
Change-Id: I0a0917eeb6fc69e2c9cbcc52a2d69434952f3554