[astro-thermal] Set Initial voltage to support Max CPU freq.
Upon boot up we were setting the voltage which would be
sufficient for 1GHz CPU freq. But when zedbooting the CPU
freq. is already set to 1.893GHz and we lower the voltage during
the init. This was causing a stall. Setting the init voltage to
support 1.893GHz fixes the bug.

ZX-2559 #done

Test: Netboot Zircon on Astro.

Change-Id: Ia7f9b06614615c39da6c5d542143809ddf2df3e3