[audio] Incorporate position modulo to mixer
This CL adds running fractional source position modulo to the core
Mix() functions, and incorporates this into the frequency response
and SINAD tests.  Those tests are the only ones that break a large
Mix pass into multiple chunks, and hence the only ones that would be
affected by this change.

In contrast to slight inaccuracies in Rate, small imprecisions in
Starting Position do not compound over time and hence do not have
as significant an impact on fidelity.

After this change, as expected, there was a small but measureable
improvement in audio fidelity, shown in updates to the threshold
values. Because the improvement is small, the work in AudioInImpl
and StandardOutputBase to adopt this feature has been deferred.

An overall reset of the Frequency Response and SINAD baseline
thresholds had not been done in some time, so the improvements
shown include the effects of more than just this CL.

Test: added multiple new audio_mixer_tests cases, manual

MTWN-128 #comment

Change-Id: I385be383e25f34711c332b815b459b6b4fc640fc