[elflib] Trim whitespace from the end of filenames in ids.txt
We hit an issue where a space at the end of a filename caused a
fatal error. This is not really the sort of thing that should
be causing symbolizer to fail at all but this is good duct tape
to solve the problem. It's probably good to do this anyway as it
increases the number of cases in which we can successfully symbolize.

Change-Id: Ie6f2e576ecb10cf53b460a23cc0489c821ed2129