[recovery_netstack] Change ArpValue to an enum
This changes ArpValue to an enum, which allows us to represent both
known values and values that are waiting on an ARP response.

It also changes the return type of the ArpTable lookup function to
return an Option<&H> instead of a ArpValue, so that users of ArpTable
do not need to interact with the ArpValue type.

NET-941

Test: None
Change-Id: I7952e209ae51056a6497715a1267e2bb98982845