[components] Pass /pkg/data path to StartupInfo
Keyed by 'data' in program_metadata, by default the component is
expected to look at the 'data' path provided in its cmx. If such field
does not exist, it is assumed to be /pkg/data/<component-name>.

Next: wire this up for flutter and dart, in topaz

TESTED=built in conjunction with
https://fuchsia-review.googlesource.com/c/topaz/+/185638
and observed that flutter mods still work.

CP-19 #comment

Change-Id: I50547d61791e84105f52511e2bc0811e9a7f4148