[fsl] Guarantee DeviceWatcher never calls Idle callback twice
The DeviceWatcher has an Idle callback that can be set
using CreateWithIdleCallback.  The Idle callback indicates
that the DeviceWatcher has finished signalling all existing
devices.  As such, the Idle callback should only be called
once.

This change guarantees that the Idle callback will only be
called once by reseting the callback once it is called, and
documents the change in the header.

TEST = CQ

Change-Id: Iea0fb2da32469dee4a2cd5bda34662012c270b45