[amlogic-video] Allow cancelling parsing
Split the wait for parsing to complete into a separate method, so the
caller can cancel parsing if it's waiting too long to finish or if it
wants to throw away the current data and seek to somewhere else.

Also repeatedly use the same parser input buffer instead of throwing it
away after every parsed chunk.

Test: amlogic_decoder_tests=true

Change-Id: I82f7629262f95d65652f06cf1a34e12436cec649