[hypervisor][time] Add support for KVM CLOCK_PAIRING hypercall
The hypercall allows to provide the guest with mapping of tsc to
wall time at some moment in time, so in the way it's the same as
boot time provided by pvclock.

Hypercall itself is not particularly interesting, but linux calls
it when kvmclock based PTP clock is enabled in the guest, which
happens for default debian_guest kernel for example. However, since
hypervisor doesn't handle VMCALL exits it fails. In general we
should not stop hypervisor on VMCALL VM exits, instead we should
return error to the guest indicating that hypercall is not supported.

Test: buildall, manually run debian_guest on Intel NUC
MAC-158 #comment In Progress

Change-Id: I79c28df80a23df6bcda8737305e6a6b169b67d78