[docs] Consolidate build and pave workflow documentation
Refactor the documentation on building and paving to put the basic
workflow in a separate quickstart doc. Rename the full
paving workflow doc to better match its location in the docs.
Rewrite clients to point to either the full doc or the quickstart,
depending on context.

Also, pull the Google-specific setup instructions out of the
getting started guide, since those instructions are already
captured in Google-internal documentation.

Test: Verbal discussion
Change-Id: I24401c01e5a9e95978e9fb17a8c3d93ddc2121d1