[audio] Profile 'int24' performance
The audio_mixer_tests binary contains functionality to performance-
profile the core Mix and ProduceOutput functions of our mixer core.
A broad array of possible combinatorial permutations is used, to
isolate the additional cost of factors such as accumulating (versus
overwriting), frame-rate-conversion (we have multiple types),
clamping, and numerical format translation to-and-from our
internally-used float32 format.

With this CL, we can now profile the '24-in-32-integer' format (in
addition to 8-bit-unsigned, 16-bit and 32-bit-float). This includes
both the intake/mixing of 24-in-32 sources as well as the emitting
of final mixed audio to 24-in-32 destinations. Somewhat
unsurprisingly, performance is slightly worse than 16-bit (and
certainly worse than float32), but the delta varies across the
various factors.

Having added these new cases, the CL also reduces the number of
repetitions in a profiling run, to preserve a soft commitment to
keep the overall perf-profiling running time to 3 minutes or less,
on both NUC and VIM2 (release builds).

Test: automated (audio_mixer_tests --profile)

Change-Id: Iac89630d008b3e888344ab646032203cdb2de75d