[vim][display] Update added/removed logic
Some of the logic surrounding displays being added and displays being
removed recently got updated.  A local variable which had been a
uint64_t became a bool, which was causing an assert in the audio code
to trip when re-plugging displays.

Update the logic to match the new boolean nature of the local
variable.

Test: manual on vim2 with an Eyoyo display, power cycling the display
      multiple times.
Change-Id: I3322f354958dd98c4875812d47ab4982f730be0e