[fidl] Prevent duplicate attributes.
Attributes can annote various elements such as method, interfaces,
types, libraries. The intent is for attributes to be a key/value pair.
Until this change however, we allowed attributes' keys to be duplicated,
potentially leading to unknown behavior in downstream libraries (e.g.
code generators).

This change enforces this requirement, and unit tests it.

We're adding an ASSERT_STR_STR to the unittesting library to check for
string containement, so as to make the tests easier to read.

Test: make tools HOST_USE_ASAN=true &&
./build-x64/host_tests/fidl-compiler-test, and full builds
Change-Id: Icc92cacc68761bd5d17d77e441300aa860e9f877

Change-Id: I5e3ec76ce2041cbc4389252b935fd5f081a23399