[fidl][fuchsia.guest] Update for FIDL readability
This mostly drops the use of redundant leading 'Guest' tag on interface
and struct names, but also changes what was the GuestManager and
GuestEnvironment interfaces to have more appropriate names. The prefix
'Instance' is used for some names to replace the removed 'Guest' to aid
clarity where it might be unclear what is being referred to.

GuestManager changes as whilst it may be the 'top level' interface provided
by the 'guestmgr' application, this interface is actually presenting a way to
manage the environments, and so is renamed to EnvironmentManager accordingly.

GuestEnvironment does not suggest an action and so is changed to
EnvironmentController.

There are also other minor method and constant name changes.

Tests: Ran guest_integration_tests, guestmgr_tests, machina_tests and manually spun up zircon and linux guests
Change-Id: Ib1577d1d9b829c75355003bacae287c78b622e91