[disk-pave] Pave empty disks by binding GPT driver manually.
Semi-recent changes to block_rr_part result in calls to block devices no
longer triggering device re-enumeration. As a consequence, when invoked
from the paver to reinitialize the GPT, the device is not detected by
the block watcher, and the GPT driver is never actually bound to the
target device.

To remedy this issue, manually take the role of the block watcher,
and bind the device from within the "InitializeGpt" function of the
paver.

ZX-2574 #done
Test: $ dd if=/dev/zero of=/dev/class/block/000 count=1024 , reboot, pave.

Change-Id: Ie30c5ed5acedf51a27bc06b0adf3df374bead1d6