[debugger] Add frame fingerprints.
Frame fingerprints are a new way to refer to a stack frame without
keeping the Frame pointer around (it will be destroyed when the thread
is run again). They have a consistent address that is valid across all
code in a function (see frame_fingerprint.h.

This implements "finish" (and by extension "until") in terms of these
frame IDs. This change should be transparent and the functionality is
unchanged, but it will make future step controllers easer to write.

TEST=unit, manual

Change-Id: I7182bfd4ef7b7bfc29b9b9831231ef1fc9bdcc08