[sysinfo] Add ioctl_sysinfo_get_board_name()
This returns the board name from the bootloader on arm64 platforms.
On x86 it the board name defaults to "pc", but a different board name
can be returned if the x86 bootloader adds a zbi_platform_id item
to the ZBI passed to the kernel.

Also fixed a problem with device_publish_metadata() not working for
devices that already exist in the device tree.

TEST: Added test to existing sysinfo test.

ZX-1775 #done

Change-Id: Ib53fcbae49656a11df59054acde0521eb64f007f