[services] Migrate garnet to new CreateNestedEnvironment API
This change migrates all users of CreateNestedEnvironment in garnet
to use |additional_services| instead of |host_directory|.

As part of this change, it was necessary to add a Launch() API to
EnclosingEnvironment. The client must call Launch() to actually start
the environment, after they have added any services to it.

CP-124 #comment

TESTED=updated EnclosingEnvironment to use new API. all garnet tests
       pass.

Change-Id: I97d1855d622828b731c3f469137eb309ee94139d